+++ crates/pathfinder/src/bin/pathfinder/main.rs
@@ -21,6 +21,7 @@ use pathfinder_storage::Storage;
 use primitive_types::H160;
 use starknet_gateway_client::GatewayApi;
 use tokio::signal::unix::{signal, SignalKind};
+use tokio::task::JoinError;
 use tracing::{info, warn};
 
 use crate::config::{NetworkConfig, StateTries};
@@ -40,10 +41,13 @@ fn main() -> anyhow::Result<()> {
         .thread_stack_size(8 * 1024 * 1024)
         .build()
         .unwrap()
-        .block_on(async { async_main().await })
+        .block_on(async {
+            async_main().await?;
+            Ok(())
+        })
 }
 
-async fn async_main() -> anyhow::Result<()> {
+async fn async_main() -> anyhow::Result<Storage> {
     if std::env::var_os("RUST_LOG").is_none() {
         // Disable all dependency logs by default.
         std::env::set_var("RUST_LOG", "pathfinder=info");
@@ -95,24 +99,12 @@ async fn async_main() -> anyhow::Result<()> {
             .default_network()
             .context("Using default Starknet network based on Ethereum configuration")?,
     };
-
-    // Spawn monitoring if configured.
-    if let Some(address) = config.monitor_address {
-        let network_label = match &network {
-            NetworkConfig::Mainnet => "mainnet",
-            NetworkConfig::SepoliaTestnet => "testnet-sepolia",
-            NetworkConfig::SepoliaIntegration => "integration-sepolia",
-            NetworkConfig::Custom { .. } => "custom",
-        };
-        spawn_monitoring(
-            network_label,
-            address,
-            readiness.clone(),
-            sync_state.clone(),
-        )
-        .await
-        .context("Starting monitoring task")?;
-    }
+    let network_label = match &network {
+        NetworkConfig::Mainnet => "mainnet",
+        NetworkConfig::SepoliaTestnet => "testnet-sepolia",
+        NetworkConfig::SepoliaIntegration => "integration-sepolia",
+        NetworkConfig::Custom { .. } => "custom",
+    };
 
     let pathfinder_context = PathfinderContext::configure_and_proxy_check(
         network,
@@ -145,6 +137,7 @@ async fn async_main() -> anyhow::Result<()> {
                 None => None,
             })
             .migrate()?;
+
     let sync_storage = storage_manager
         // 5 is enough for normal sync operations, and then `available_parallelism` for
         // the rayon thread pool workers to use.
@@ -185,16 +178,16 @@ Hint: This is usually caused by exceeding the file descriptor limit of your syst
 Hint: This is usually caused by exceeding the file descriptor limit of your system.
       Try increasing the file limit to using `ulimit` or similar tooling.",
         )?;
-
+    // 5 is enough for normal sync operations, and then `available_parallelism` for
+    // the rayon thread pool workers to use.
     let p2p_storage = storage_manager
-        .create_pool(NonZeroU32::new(1).unwrap())
+        .create_pool(NonZeroU32::new(5 + available_parallelism.get() as u32).unwrap())
         .context(
             r"Creating database connection pool for p2p
 
 Hint: This is usually caused by exceeding the file descriptor limit of your system.
       Try increasing the file limit to using `ulimit` or similar tooling.",
         )?;
-
     info!(location=?pathfinder_context.database, "Database migrated.");
     verify_database(
         &sync_storage,
@@ -211,6 +204,12 @@ Hint: This is usually caused by exceeding the file descriptor limit of your syst
         .context(r"Creating database transaction")?
         .prune_tries()
         .context("Pruning tries on startup")?;
+
+    // Register signal handlers here, because we want to be able to interrupt long
+    // running migrations or trie pruning. No tasks are spawned before this point so
+    // we don't worry about detachment.
+    let mut term_signal = signal(SignalKind::terminate())?;
+    let mut int_signal = signal(SignalKind::interrupt())?;
 
     let (tx_pending, rx_pending) = tokio::sync::watch::channel(Default::default());
 
@@ -257,16 +256,44 @@ Hint: This is usually caused by exceeding the file descriptor limit of your syst
         None => rpc_server,
     };
 
+    // Spawn monitoring if configured.
+    if let Some(address) = config.monitor_address {
+        spawn_monitoring(
+            network_label,
+            address,
+            readiness.clone(),
+            sync_state.clone(),
+        )
+        .await
+        .context("Starting monitoring task")?;
+    }
+
+    // From this point onwards, until the final select, we don't exit the process
+    // even if some error is encountered or a signal is received as it would result
+    // in tasks being detached and cancelled abruptly without a chance to clean
+    // up. We need to wait for the final select where we can cancel all the tasks
+    // and wait for them to finish. Only then can we exit the process and return an
+    // error if some of the tasks failed or no error if we have received a signal.
+
     let (p2p_handle, gossiper, p2p_client) = start_p2p(
         pathfinder_context.network_id,
         p2p_storage,
         config.p2p.clone(),
     )
-    .await?;
+    .await
+    .unwrap_or_else(|error| {
+        (
+            tokio::task::spawn(std::future::ready(
+                Err(error.context("P2P failed to start")),
+            )),
+            Default::default(),
+            None,
+        )
+    });
 
     let sync_handle = if config.is_sync_enabled {
         start_sync(
-            sync_storage,
+            sync_storage.clone(),
             pathfinder_context,
             ethereum.client,
             sync_state.clone(),
@@ -284,59 +311,64 @@ Hint: This is usually caused by exceeding the file descriptor limit of your syst
     };
 
     let rpc_handle = if config.is_rpc_enabled {
-        let (rpc_handle, local_addr) = rpc_server
+        match rpc_server
             .with_max_connections(config.max_rpc_connections.get())
             .spawn()
             .await
-            .context("Starting the RPC server")?;
-        info!("📡 HTTP-RPC server started on: {}", local_addr);
-        rpc_handle
+        {
+            Ok((rpc_handle, on)) => {
+                info!(%on, "📡 RPC server started");
+                rpc_handle
+            }
+            Err(error) => tokio::task::spawn(std::future::ready(Err(
+                error.context("RPC server failed to start")
+            ))),
+        }
     } else {
         tokio::spawn(std::future::pending())
     };
 
     if !config.disable_version_update_check {
-        tokio::spawn(update::poll_github_for_releases());
+        util::task::spawn(update::poll_github_for_releases());
     }
-
-    let mut term_signal = signal(SignalKind::terminate())?;
-    let mut int_signal = signal(SignalKind::interrupt())?;
 
     // We are now ready.
     readiness.store(true, std::sync::atomic::Ordering::Relaxed);
 
     // Monitor our critical spawned process tasks.
-    tokio::select! {
-        result = sync_handle => {
-            match result {
-                Ok(task_result) => tracing::error!("Sync process ended unexpected with: {:?}", task_result),
-                Err(err) => tracing::error!("Sync process ended unexpected; failed to join task handle: {:?}", err),
-            }
-            anyhow::bail!("Unexpected shutdown");
-        }
-        result = rpc_handle => {
-            match result {
-                Ok(_) => tracing::error!("RPC server process ended unexpectedly"),
-                Err(err) => tracing::error!(error=%err, "RPC server process ended unexpectedly"),
-            }
-            anyhow::bail!("Unexpected shutdown");
-        }
-        result = p2p_handle => {
-            match result {
-                Ok(_) => tracing::error!("P2P process ended unexpectedly"),
-                Err(err) => tracing::error!(error=%err, "P2P process ended unexpectedly"),
-            }
-            anyhow::bail!("Unexpected shutdown");
-        }
+    let main_result = tokio::select! {
+        result = sync_handle => handle_critical_task_result("Sync", result),
+        result = rpc_handle => handle_critical_task_result("RPC", result),
+        result = p2p_handle => handle_critical_task_result("P2P", result),
         _ = term_signal.recv() => {
-            tracing::info!("TERM signal received, exiting gracefully");
+            tracing::info!("TERM signal received");
             Ok(())
         }
         _ = int_signal.recv() => {
-            tracing::info!("INT signal received, exiting gracefully");
+            tracing::info!("INT signal received");
             Ok(())
         }
+    };
+
+    // If we get here either a signal was received or a task ended unexpectedly,
+    // which means we need to cancel all the remaining tasks.
+    tracing::info!("Shutdown started, waiting for tasks to finish...");
+    util::task::tracker::close();
+    // Force exit after a grace period
+    match tokio::time::timeout(config.shutdown_grace_period, util::task::tracker::wait()).await {
+        Ok(_) => {
+            tracing::info!("Shutdown finished successfully")
+        }
+        Err(_) => {
+            tracing::error!("Some tasks failed to finish in time, forcing exit");
+        }
     }
+
+    // If a RO db connection pool remains after all RW connection pools have been
+    // dropped, WAL & SHM files are never cleaned up. To avoid this, we make sure
+    // that all RO pools and all but one RW pools are dropped when task tracker
+    // finishes waiting, and then we drop the last RW pool.
+    main_result.map(|_| sync_storage)
 }
 
 #[cfg(feature = "tokio-console")]
@@ -412,7 +444,7 @@ async fn start_p2p(
     storage: Storage,
     config: config::P2PConfig,
 ) -> anyhow::Result<(
-    tokio::task::JoinHandle<()>,
+    tokio::task::JoinHandle<anyhow::Result<()>>,
     state::Gossiper,
     Option<p2p::client::peer_agnostic::Client>,
 )> {
@@ -496,7 +528,7 @@ async fn start_p2p(
     _: Storage,
     _: config::P2PConfig,
 ) -> anyhow::Result<(
-    tokio::task::JoinHandle<()>,
+    tokio::task::JoinHandle<anyhow::Result<()>>,
     state::Gossiper,
     Option<p2p::client::peer_agnostic::Client>,
 )> {
@@ -614,7 +646,7 @@ fn start_feeder_gateway_sync(
         fetch_casm_from_fgw: config.fetch_casm_from_fgw,
     };
 
-    tokio::spawn(state::sync(sync_context, state::l1::sync, state::l2::sync))
+    util::task::spawn(state::sync(sync_context, state::l1::sync, state::l2::sync))
 }
 
 #[cfg(feature = "p2p")]
@@ -641,7 +673,7 @@ fn start_p2p_sync(
         verify_tree_hashes,
         block_hash_db: Some(BlockHashDb::new(pathfinder_context.network)),
     };
-    tokio::spawn(sync.run())
+    util::task::spawn(sync.run())
 }
 
 /// Spawns the monitoring task at the given address.
@@ -875,16 +907,14 @@ async fn verify_database(
     gateway_client: &starknet_gateway_client::Client,
 ) -> anyhow::Result<()> {
     let storage = storage.clone();
-    let db_genesis = tokio::task::spawn_blocking(move || {
-        let mut conn = storage.connection().context("Create database connection")?;
-        let tx = conn.transaction().context("Create database transaction")?;
 
-        tx.block_id(BlockNumber::GENESIS.into())
-    })
-    .await
-    .context("Joining database task")?
-    .context("Fetching genesis hash from database")?
-    .map(|x| x.1);
+    let mut conn = storage.connection().context("Create database connection")?;
+    let tx = conn.transaction().context("Create database transaction")?;
+
+    let db_genesis = tx
+        .block_id(BlockNumber::GENESIS.into())
+        .context("Fetching genesis hash from database")?
+        .map(|x| x.1);
 
     if let Some(database_genesis) = db_genesis {
         use pathfinder_common::consts::{
@@ -925,4 +955,29 @@ async fn verify_database(
     }
 
     Ok(())
+}
+
+fn handle_critical_task_result(
+    task_name: &str,
+    task_result: Result<anyhow::Result<()>, JoinError>,
+) -> anyhow::Result<()> {
+    match task_result {
+        Ok(task_result) => {
+            tracing::error!(?task_result, "{} task ended unexpectedly", task_name);
+            task_result
+        }
+        Err(error) if error.is_panic() => {
+            tracing::error!(%error, "{} task panicked", task_name);
+            Err(anyhow::anyhow!("{} task panicked", task_name))
+        }
+        // Cancelling all tracked tasks via [`util::task::tracker::close()`] does not cause join
+        // errors on registered task handles, so this is unexpected and we should threat it as error
+        Err(_) => {
+            tracing::error!("{} task was cancelled unexpectedly", task_name);
+            Err(anyhow::anyhow!(
+                "{} task was cancelled unexpectedly",
+                task_name
+            ))
+        }
+    }
 }