+++ crates/pathfinder/src/sync/track.rs
@@ -206,7 +206,7 @@ impl<L, P> HeaderSource<L, P> {
             mut start,
         } = self;
 
-        tokio::spawn(async move {
+        util::task::spawn(async move {
             let mut latest_onchain = Box::pin(latest_onchain);
             while let Some(latest_onchain) = latest_onchain.next().await {
                 let mut headers =
@@ -241,7 +241,7 @@ impl StateDiffFanout {
         let (d1_tx, d1_rx) = tokio::sync::mpsc::channel(buffer);
         let (d2_tx, d2_rx) = tokio::sync::mpsc::channel(buffer);
 
-        tokio::spawn(async move {
+        util::task::spawn(async move {
             while let Some(state_update) = source.recv().await {
                 let is_err = state_update.is_err();
 
@@ -288,7 +288,7 @@ impl TransactionsFanout {
         let (t_tx, t_rx) = tokio::sync::mpsc::channel(buffer);
         let (e_tx, e_rx) = tokio::sync::mpsc::channel(buffer);
 
-        tokio::spawn(async move {
+        util::task::spawn(async move {
             while let Some(transactions) = source.recv().await {
                 let is_err = transactions.is_err();
 
@@ -329,7 +329,7 @@ impl HeaderFanout {
         let (s_tx, s_rx) = tokio::sync::mpsc::channel(buffer);
         let (t_tx, t_rx) = tokio::sync::mpsc::channel(buffer);
 
-        tokio::spawn(async move {
+        util::task::spawn(async move {
             while let Some(signed_header) = source.recv().await {
                 let is_err = signed_header.is_err();
 
@@ -381,7 +381,8 @@ impl<P> TransactionSource<P> {
         P: Clone + BlockClient + Send + 'static,
     {
         let (tx, rx) = tokio::sync::mpsc::channel(1);
-        tokio::spawn(async move {
+
+        util::task::spawn(async move {
             let Self { p2p, mut headers } = self;
 
             while let Some(header) = headers.next().await {
@@ -456,7 +457,8 @@ impl<P> EventSource<P> {
         P: Clone + BlockClient + Send + 'static,
     {
         let (tx, rx) = tokio::sync::mpsc::channel(1);
-        tokio::spawn(async move {
+
+        util::task::spawn(async move {
             let Self {
                 p2p,
                 mut transactions,
@@ -539,7 +541,8 @@ impl<P> StateDiffSource<P> {
         P: Clone + BlockClient + Send + 'static,
     {
         let (tx, rx) = tokio::sync::mpsc::channel(1);
-        tokio::spawn(async move {
+
+        util::task::spawn(async move {
             let Self { p2p, mut headers } = self;
 
             while let Some(header) = headers.next().await {
@@ -595,7 +598,8 @@ impl<P> ClassSource<P> {
         P: Clone + BlockClient + Send + 'static,
     {
         let (tx, rx) = tokio::sync::mpsc::channel(1);
-        tokio::spawn(async move {
+
+        util::task::spawn(async move {
             let Self {
                 p2p,
                 mut declarations,
@@ -663,7 +667,7 @@ impl BlockStream {
     fn spawn(mut self) -> SyncReceiver<BlockData> {
         let (tx, rx) = tokio::sync::mpsc::channel(1);
 
-        tokio::spawn(async move {
+        util::task::spawn(async move {
             loop {
                 let Some(result) = self.next().await else {
                     return;