+++ crates/rpc/src/jsonrpc/router/subscription.rs
@@ -176,7 +176,7 @@ where
                 let first_block = pathfinder_storage::BlockId::try_from(first_block)
                     .map_err(|e| RpcError::InvalidParams(e.to_string()))?;
                 let storage = router.context.storage.clone();
-                let current_block = tokio::task::spawn_blocking(move || -> Result<_, RpcError> {
+                let current_block = util::task::spawn_blocking(move |_| -> Result<_, RpcError> {
                     let mut conn = storage.connection().map_err(RpcError::InternalError)?;
                     let db = conn.transaction().map_err(RpcError::InternalError)?;
                     db.block_number(first_block)
@@ -189,7 +189,7 @@ where
             }
         };
 
-        Ok(tokio::spawn(async move {
+        Ok(util::task::spawn(async move {
             let _subscription_guard = SubscriptionsGuard {
                 subscription_id,
                 subscriptions,
@@ -244,7 +244,7 @@ where
 
             // Subscribe to new blocks. Receive the first subscription message.
             let (tx1, mut rx1) = mpsc::channel::<SubscriptionMessage<T::Notification>>(1024);
-            tokio::spawn({
+            util::task::spawn({
                 let params = params.clone();
                 let context = router.context.clone();
                 let tx = tx.clone();
@@ -352,7 +352,7 @@ pub fn split_ws(ws: WebSocket, version: RpcVersion) -> (WsSender, WsReceiver) {
     let (mut ws_sender, mut ws_receiver) = ws.split();
     // Send messages to the websocket using an MPSC channel.
     let (sender_tx, mut sender_rx) = mpsc::channel::<Result<Message, RpcResponse>>(1024);
-    tokio::spawn(async move {
+    util::task::spawn(async move {
         while let Some(msg) = sender_rx.recv().await {
             match msg {
                 Ok(msg) => {
@@ -379,7 +379,7 @@ pub fn split_ws(ws: WebSocket, version: RpcVersion) -> (WsSender, WsReceiver) {
     });
     // Receive messages from the websocket using an MPSC channel.
     let (receiver_tx, receiver_rx) = mpsc::channel::<Result<Message, axum::Error>>(1024);
-    tokio::spawn(async move {
+    util::task::spawn(async move {
         while let Some(msg) = ws_receiver.next().await {
             if receiver_tx.send(msg).await.is_err() {
                 break;
@@ -397,7 +397,7 @@ pub fn handle_json_rpc_socket(
     let subscriptions: Arc<DashMap<SubscriptionId, tokio::task::JoinHandle<()>>> =
         Default::default();
     // Read and handle messages from the websocket.
-    tokio::spawn(async move {
+    util::task::spawn(async move {
         loop {
             let request = match ws_rx.recv().await {
                 Some(Ok(Message::Text(msg))) => msg,