+++ crates/rpc/src/method/simulate_transactions.rs
@@ -34,7 +34,7 @@ pub async fn simulate_transactions(
     input: SimulateTransactionInput,
 ) -> Result<Output, SimulateTransactionError> {
     let span = tracing::Span::current();
-    tokio::task::spawn_blocking(move || {
+    util::task::spawn_blocking(move |_| {
         let _g = span.enter();
 
         let skip_validate = input