+++ crates/rpc/src/method/trace_block_transactions.rs
@@ -46,7 +46,7 @@ pub async fn trace_block_transactions(
     let span = tracing::Span::current();
 
     let storage = context.execution_storage.clone();
-    let traces = tokio::task::spawn_blocking(move || {
+    let traces = util::task::spawn_blocking(move |_| {
         let _g = span.enter();
 
         let mut db = storage.connection()?;